# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Amrutam Telemedicine API",
    "description": "Production-grade telemedicine backend: auth/MFA, doctor availability & booking (idempotent, saga-driven), consultations, prescriptions, payments, admin analytics.",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "def-0": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {},
              "requestId": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      }
    }
  },
  "paths": {
...
    "/doctors": {
      "get": {
        "summary": "Search & filter doctors (cached, keyset-paginated)",
        "tags": [
          "doctors"
        ],
        "parameters": [
          {
            "schema": {
              "maxLength": 100,
              "type": "string"
            },
            "in": "query",
            "name": "q",
            "required": false
          },
          {
            "schema": {
              "maxLength": 60,
              "type": "string"
            },
            "in": "query",
            "name": "specialization",
            "required": false
          },
          {
            "schema": {
              "minimum": 0,
              "maximum": 5,
              "type": "number"
            },
            "in": "query",
            "name": "minRating",
            "required": false
          },
          {
            "schema": {
              "minimum": 0,
              "type": "number"
            },
            "in": "query",
            "name": "maxFee",
            "required": false
          },
          {
            "schema": {
              "maxLength": 40,
              "type": "string"
            },
            "in": "query",
            "name": "language",
            "required": false
          },
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "enum": [
                    "rating"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "fee"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "experience"
                  ]
                }
              ]
            },
            "in": "query",
            "name": "sort",
            "required": false
          },
          {
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "integer"
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "cursor",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "format": "uuid",
                            "type": "string"
                          },
                          "userId": {
                            "format": "uuid",
                            "type": "string"
                          },
                          "displayName": {
                            "type": "string"
                          },
                          "bio": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "yearsExperience": {
                            "type": "integer"
                          },
                          "consultationFee": {
                            "type": "number"
                          },
                          "currency": {
                            "type": "string"
                          },
                          "languages": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "ratingAvg": {
                            "type": "number"
                          },
                          "ratingCount": {
                            "type": "integer"
                          },
                          "isVerified": {
                            "type": "boolean"
                          },
                          "isAccepting": {
                            "type": "boolean"
                          },
                          "timezone": {
                            "type": "string"
                          },
                          "specializations": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "id": {
                                  "type": "integer"
                                },
                                "slug": {
                                  "type": "string"
                                },
                                "name": {
                                  "type": "string"
                                }
                              },
                              "required": [
                                "id",
                                "slug",
                                "name"
                              ]
                            }
                          }
                        },
                        "required": [
                          "id",
                          "userId",
                          "displayName",
                          "bio",
                          "yearsExperience",
                          "consultationFee",
                          "currency",
                          "languages",
                          "ratingAvg",
                          "ratingCount",
                          "isVerified",
                          "isAccepting",
                          "timezone",
                          "specializations"
                        ]
                      }
                    },
                    "nextCursor": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  },
                  "required": [
                    "items",
                    "nextCursor"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          }
        }
      }
    },
    "/doctors/{id}": {
      "get": {
        "summary": "Fetch a doctor profile (cache-aside)",
        "tags": [
          "doctors"
        ],
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "userId": {
                      "format": "uuid",
                      "type": "string"
                    },
                    "displayName": {
                      "type": "string"
                    },
                    "bio": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "yearsExperience": {
                      "type": "integer"
                    },
                    "consultationFee": {
                      "type": "number"
                    },
                    "currency": {
                      "type": "string"
                    },
                    "languages": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "ratingAvg": {
                      "type": "number"
                    },
                    "ratingCount": {
                      "type": "integer"
                    },
                    "isVerified": {
                      "type": "boolean"
                    },
                    "isAccepting": {
                      "type": "boolean"
                    },
                    "timezone": {
                      "type": "string"
                    },
                    "specializations": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "slug": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "slug",
                          "name"
                        ]
                      }
                    }
                  },
                  "required": [
                    "id",
                    "userId",
                    "displayName",
                    "bio",
                    "yearsExperience",
                    "consultationFee",
                    "currency",
                    "languages",
                    "ratingAvg",
                    "ratingCount",
                    "isVerified",
                    "isAccepting",
                    "timezone",
                    "specializations"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "429": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
}